+++ pkg/network/node.go
@@ -16,7 +16,7 @@ import (
 
 func NewHiddenLakeNode(
 	pName string,
-	pCfg ISettings,
+	pSettings ISettings,
 	pPrivKey asymmetric.IPrivKey,
 	pKVDatabase database.IKVDatabase,
 ) anonymity.INode {
@@ -24,21 +24,21 @@ func NewHiddenLakeNode(
 		anonymity.NewSettings(&anonymity.SSettings{
 			FServiceName:  pName,
 			FNetworkMask:  hiddenlake.GSettings.FProtoMask.FNetwork,
-			FFetchTimeout: pCfg.GetFetchTimeout(),
+			FFetchTimeout: pSettings.GetFetchTimeout(),
 		}),
 		// Insecure to use logging in real anonymity projects!
 		// Logging should only be used in overview or testing;
-		pCfg.GetLogger(),
+		pSettings.GetLogger(),
 		pKVDatabase,
 		network.NewNode(
 			network.NewSettings(&network.SSettings{
-				FAddress:      pCfg.GetTCPAddress(),
+				FAddress:      pSettings.GetTCPAddress(),
 				FMaxConnects:  hiddenlake.GSettings.FNetworkManager.FConnectsLimiter,
 				FReadTimeout:  hiddenlake.GSettings.GetReadTimeout(),
 				FWriteTimeout: hiddenlake.GSettings.GetWriteTimeout(),
 				FConnSettings: conn.NewSettings(&conn.SSettings{
-					FMessageSettings:       pCfg.GetMessageSettings(),
-					FLimitMessageSizeBytes: pCfg.GetMessageSizeBytes(),
+					FMessageSettings:       pSettings.GetMessageSettings(),
+					FLimitMessageSizeBytes: pSettings.GetMessageSizeBytes(),
 					FWaitReadTimeout:       hiddenlake.GSettings.GetWaitTimeout(),
 					FDialTimeout:           hiddenlake.GSettings.GetDialTimeout(),
 					FReadTimeout:           hiddenlake.GSettings.GetReadTimeout(),
@@ -50,18 +50,18 @@ func NewHiddenLakeNode(
 		queue.NewQBProblemProcessor(
 			queue.NewSettings(&queue.SSettings{
 				FMessageConstructSettings: net_message.NewConstructSettings(&net_message.SConstructSettings{
-					FSettings: pCfg.GetMessageSettings(),
-					FParallel: pCfg.GetParallel(),
+					FSettings: pSettings.GetMessageSettings(),
+					FParallel: pSettings.GetParallel(),
 				}),
 				FNetworkMask: hiddenlake.GSettings.FProtoMask.FNetwork,
-				FQueuePeriod: pCfg.GetQueuePeriod(),
+				FQueuePeriod: pSettings.GetQueuePeriod(),
 				FPoolCapacity: [2]uint64{
 					hiddenlake.GSettings.FQueueCapacity.FMain,
 					hiddenlake.GSettings.FQueueCapacity.FRand,
 				},
 			}),
 			func() client.IClient {
-				client := client.NewClient(pPrivKey, pCfg.GetMessageSizeBytes())
+				client := client.NewClient(pPrivKey, pSettings.GetMessageSizeBytes())
 				if client.GetPayloadLimit() <= encoding.CSizeUint64 {
 					panic(`client.GetPayloadLimit() <= encoding.CSizeUint64`)
 				}